requirements.txt
python-telegram-bot
requests
aiohttp
fastapi
uvicorn
pydantic
python-multipart
sqlite3
